pyproject.toml
[project]
name = "rqe-mappo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gymnasium>=1.2.2",
    "matplotlib>=3.10.7",
    "pettingzoo[mpe]>=1.25.0",
    "pillow>=12.0.0",
    "seaborn>=0.13.2",
    "sumo-rl>=1.4.5",
    "supersuit>=3.10.0",
    "tensorboard>=2.20.0",
    "torch>=2.9.0",
    "torchaudio>=2.9.0",
    "torchvision>=0.24.0",
]
